# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exactamente API",
    "description": "Recursos universitarios de Exactas: parciales, resúmenes y finales.\n\nEste documento se genera desde los schemas Zod del backend, así que no\npuede desincronizarse de la validación. Los clientes (web, panel admin y\nservidor MCP) generan sus tipos desde acá.\n\n**Paginación:** `page` (default 1) y `limit` (default 20, máx 100).\nExcepción: `GET /careers` devuelve `{ data: [...] }` sin paginar.\n\n**Auth:** las rutas protegidas piden `Authorization: Bearer <jwt>`.\nLas de `/admin/*` requieren rol `admin` o `superadmin`.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://github.com/exactamente-ar/exactamente-backend"
    }
  },
  "servers": [
    {
      "url": "https://api.exactamente.com.ar/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "bearer",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SubjectCareer": {
        "id": "SubjectCareer",
        "type": "object",
        "properties": {
          "careerId": {
            "type": "string"
          },
          "careerName": {
            "type": "string"
          },
          "facultyId": {
            "type": "string"
          },
          "facultyName": {
            "type": "string"
          },
          "universityId": {
            "type": "string"
          },
          "universityName": {
            "type": "string"
          },
          "planId": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "quadmester": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "careerId",
          "careerName",
          "facultyId",
          "facultyName",
          "universityId",
          "universityName",
          "planId",
          "year",
          "quadmester"
        ],
        "additionalProperties": false
      },
...
      "SubjectDetail": {
        "id": "SubjectDetail",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "facultyId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "description": "Se regenera si cambia el title",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "urlMoodle": {
            "description": "Puede ser cadena vacía, nunca null",
            "type": "string"
          },
          "urlPrograma": {
            "description": "Puede ser cadena vacía, nunca null",
            "type": "string"
          },
          "year": {
            "description": "Año de cursada en el plan",
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "quadmester": {
            "description": "1 o 2",
            "type": "integer",
            "minimum": 1,
            "maximum": 2
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "careers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectCareer"
            }
          }
        },
        "required": [
          "id",
          "facultyId",
          "title",
          "slug",
          "description",
          "urlMoodle",
          "urlPrograma",
          "year",
          "quadmester",
          "createdAt",
          "updatedAt",
          "careers"
        ],
        "additionalProperties": false
      },
      "AdminSubject": {
        "id": "AdminSubject",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "facultyId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "description": "Se regenera si cambia el title",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "urlMoodle": {
            "description": "Puede ser cadena vacía, nunca null",
            "type": "string"
          },
          "urlPrograma": {
            "description": "Puede ser cadena vacía, nunca null",
            "type": "string"
          },
          "year": {
            "description": "Año de cursada en el plan",
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "quadmester": {
            "description": "1 o 2",
            "type": "integer",
            "minimum": 1,
            "maximum": 2
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "resourceCount": {
            "description": "Recursos en estado published",
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "facultyId",
          "title",
          "slug",
          "description",
          "urlMoodle",
          "urlPrograma",
          "year",
          "quadmester",
          "createdAt",
          "updatedAt",
          "resourceCount"
        ],
        "additionalProperties": false
      },
      "CareerPlanSubject": {
        "id": "CareerPlanSubject",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "facultyId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "urlMoodle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "urlPrograma": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "year": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "quadmester": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "resourceCount": {
            "description": "Recursos en estado published",
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "facultyId",
          "title",
          "description",
          "urlMoodle",
          "urlPrograma",
          "year",
          "quadmester",
          "createdAt",
          "resourceCount"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/subjects/{id}": {
      "get": {
        "operationId": "getApiV1SubjectsById",
        "tags": [
          "Materias"
        ],
        "summary": "Detalle de una materia",
        "description": "A diferencia del listado, NO incluye `resourceCounts`.",
        "responses": {
          "200": {
            "description": "Materia con sus carreras",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subject": {
                      "$ref": "#/components/schemas/SubjectDetail"
                    }
                  },
                  "required": [
                    "subject"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
...
    "/api/v1/admin/career-plans/{id}/subjects": {
      "get": {
        "operationId": "getApiV1AdminCareerPlansByIdSubjects",
        "tags": [
          "Admin"
        ],
        "summary": "Materias de un plan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Materias del plan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CareerPlanSubject"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "page": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "totalPages": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
...
    "/api/v1/admin/subjects": {
      "get": {
        "operationId": "getApiV1AdminSubjects",
        "tags": [
          "Admin"
        ],
        "summary": "Listar materias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminSubject"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "page": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "totalPages": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "totalPages"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postApiV1AdminSubjects",
        "tags": [
          "Admin"
        ],
        "summary": "Crear materia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Materia creada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "facultyId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "slug": {
                      "description": "Se regenera si cambia el title",
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "urlMoodle": {
                      "description": "Puede ser cadena vacía, nunca null",
                      "type": "string"
                    },
                    "urlPrograma": {
                      "description": "Puede ser cadena vacía, nunca null",
                      "type": "string"
                    },
                    "year": {
                      "description": "Año de cursada en el plan",
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 5
                    },
                    "quadmester": {
                      "description": "1 o 2",
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 2
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "facultyId",
                    "title",
                    "slug",
                    "description",
                    "urlMoodle",
                    "urlPrograma",
                    "year",
                    "quadmester",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}